refactor: split background/index.ts into focused modules
Extract the 509-line monolith into:
- config.ts: debug function, getApiUrl, shared constants
- candidate-sync.ts: sendCandidates, buffer management, storeRecentCandidates
- credential-sync.ts: sendCredentialsToServer, processPendingCredentials
- schedule-capture.ts: sendScheduleCapture
- message-handler.ts: chrome.runtime.onMessage dispatcher
- index.ts: slim orchestrator (register listeners, start intervals)

Pure refactoring — no behavior changes.
